requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
idna==2.5
numpy==1.13.0
pandas==0.20.2
pyPdf==1.13
python-dateutil==2.6.0
pytz==2017.2
requests==2.17.3
six==1.10.0
tabula-py==0.9.0
urllib3==1.21.1
